selected headers moved to public include folder
