+++ Code/Include/ROS2/Frame/NamespaceConfiguration.h
