+++ Code/Source/Camera/CameraSensor.h
@@ -7,7 +7,7 @@
  */
 #pragma once
 
-#include "ROS2GemUtilities.h"
+#include "ROS2/ROS2GemUtilities.h"
 #include <Atom/Feature/Utils/FrameCaptureBus.h>
 #include <chrono>
