+++ Code/Source/GNSS/ROS2GNSSSensorComponent.cpp
@@ -7,13 +7,13 @@
  */
 
 #include "GNSS/ROS2GNSSSensorComponent.h"
-#include "AzCore/Math/Matrix4x4.h"
-#include "Frame/ROS2FrameComponent.h"
+#include "ROS2/Frame/ROS2FrameComponent.h"
 #include "ROS2/ROS2Bus.h"
-#include "ROS2GemUtilities.h"
+#include "ROS2/Utilities/ROS2Names.h"
+#include "ROS2/ROS2GemUtilities.h"
 #include "Utilities/ROS2Conversions.h"
-#include "Utilities/ROS2Names.h"
 
+#include <AzCore/Math/Matrix4x4.h>
 #include <AzCore/Serialization/EditContext.h>
 #include <AzCore/Serialization/EditContextConstants.inl>
