+++ Code/Source/Lidar/ROS2LidarSensorComponent.cpp
@@ -7,10 +7,10 @@
  */
 
 #include "Lidar/ROS2LidarSensorComponent.h"
-#include "Frame/ROS2FrameComponent.h"
+#include "ROS2/Frame/ROS2FrameComponent.h"
 #include "Lidar/LidarTemplateUtils.h"
 #include "ROS2/ROS2Bus.h"
-#include "Utilities/ROS2Names.h"
+#include "ROS2/Utilities/ROS2Names.h"
 #include <Atom/RPI.Public/AuxGeom/AuxGeomFeatureProcessorInterface.h>
 #include <Atom/RPI.Public/RPISystemInterface.h>
 #include <Atom/RPI.Public/Scene.h>
