+++ Code/Source/ROS2GemUtilities.cpp
@@ -6,7 +6,7 @@
  *
  */
 
-#include "ROS2GemUtilities.h"
+#include "../Include/ROS2/ROS2GemUtilities.h"
 #include <AzCore/std/string/regex.h>
 #include <AzToolsFramework/API/EntityCompositionRequestBus.h>
