+++ Code/Source/RobotControl/ControlConfiguration.cpp
@@ -7,7 +7,7 @@
  */
 
 #include "RobotControl/ControlConfiguration.h"
-#include "Utilities/ROS2Names.h"
+#include "ROS2/Utilities/ROS2Names.h"
 #include <AzCore/Serialization/EditContext.h>
 #include <AzCore/Serialization/EditContextConstants.inl>
