+++ Code/Source/RobotImporter/URDF/URDFPrefabMaker.cpp
@@ -7,8 +7,8 @@
  */
 
 #include "RobotImporter/URDF/URDFPrefabMaker.h"
-#include "Frame/ROS2FrameComponent.h"
-#include "ROS2GemUtilities.h"
+#include "ROS2/Frame/ROS2FrameComponent.h"
+#include "ROS2/ROS2GemUtilities.h"
 #include "RobotControl/ROS2RobotControlComponent.h"
 #include "RobotImporter/URDF/CollidersMaker.h"
 #include "RobotImporter/URDF/PrefabMakerUtils.h"
