+++ Code/Source/Sensor/PublisherConfiguration.cpp
@@ -7,7 +7,7 @@
  */
 
 #include "PublisherConfiguration.h"
-#include "Utilities/ROS2Names.h"
+#include "ROS2/Utilities/ROS2Names.h"
 #include <AzCore/Serialization/EditContext.h>
 #include <AzCore/Serialization/SerializeContext.h>
