+++ Code/Source/Sensor/ROS2SensorComponent.h
@@ -7,7 +7,7 @@
  */
 #pragma once
 
-#include "ROS2GemUtilities.h"
+#include "ROS2/ROS2GemUtilities.h"
 #include "SensorConfiguration.h"
 #include <AzCore/Component/Component.h>
 #include <AzCore/Component/TickBus.h>
