+++ Code/ros2_header_files.cmake
@@ -4,5 +4,10 @@
 # SPDX-License-Identifier: Apache-2.0 OR MIT
 
 set(FILES
-    Include/ROS2/ROS2Bus.h
-)
+        Include/ROS2/Frame/NamespaceConfiguration.h
+        Include/ROS2/Frame/ROS2FrameComponent.h
+        Include/ROS2/Frame/ROS2Transform.h
+        Include/ROS2/ROS2Bus.h
+        Include/ROS2/ROS2GemUtilities.h
+        Include/ROS2/Utilities/ROS2Names.h
+        )
